update README
The position of a moving planet towards to another fixed planet. -> The trace of a moving planet towards to another fixed planet.
(It is better to use trace instead of position because all the following sentences illustrate the things about the track of moving planets. To avoid using "track" several times, I suggest to use "trace" here.)

In this diagram, the smallest or the inner
ellipse is the original track of this moving planet. With time
increasing, this diagram shows the moving track of this planet starts
expanding to right side slowly but the track is still ellipse. In the
other world, the shape of moving track is always ellipse whose
center moving to left side slowing and Semi-major axis increases
slowly as well. The right side of the moving track does not have
great change.
-> In this diagram, the inner ellipse is the original track of the moving planet. This diagram shows that the trace starts expanding to right side slowly and the trace is still an ellipse. The center of the track moves to the left side slightly and semi-major axis increases slowly. The right side of the track does not have great changes.
(remove some redundancy and reconstruct some sentences)